Instead of deployment, convert the code focused on binaries to be 'packages'.
This now requires two-level commands, e.g. admin package init and
admin package build, so the utilities for parsing and getting help
with commands needed to be modified to handle up to two command
parts.

The layout of data will be split between packages, for binary builds,
and deployed services, like a space server that you want to run out of
a package.

The rest of the changes are mainly just shifting things around to make
them work with this new layout.