Add basic service management: init, start, stop, destroy.
Add service management. Services work in a separate namespace from
packages. Initializing a service means specifying a package to run
from, optionally copying in a template for that service, and setting
up a minimal configuration to be modified by the user. Start and stop
use start-stop-daemon to start and stop the service. Destroy makes
sure the service isn't running and completely destroys all of its
contents.