Add melville template.
The melville template is the space server and scene setup we've been
using to test melville. It doesn't actually include any of melville
because melville is client side only. It is a simple single-server
space and few objects in a db with one script to help generate the
rest of the world.