Add template configuration and template instantiation commands.
Add commands to take entire templates and instantiate and destroy all
of their services with a prefix, as well as start and stop them as a
whole. The start and stop commands have configurations that allow you
to control the order and timing of the services starting.

The configuration for templates is a simple list of services with a
few properties. We copy it into the instantiated template directory to
ensure we have a copy that doesn't change when the template is
edited. This also adds a template config for the melville template.