Add reinit commands.
Add reinit commands for services and templates. This mostly just
refactors the init code into error checking code that makes sure the
destinations are clear, used only for init, and the code that actually
generates the new services, used for both init and reinit. The only
significant code changes are to make the generation code work when
files already exist.

There's no attempt to merge, so any configuration files *will* be
overwritten and customizations lost. This is just a way to force an
existing instantiation back to a working state or to match it to an
updated template.