Add monit support.
Add support for running services and templates under monit. The
template functionality just builds on the service interaction with
monit. For services, we modify the commands so the previous start/stop
commands, which use start-stop-daemon, are now rawstart/rawstop. The
new start and stop commands are wrappers which decide whether to call
the raw commands or a monit command based on a flag in the service's
configuration.

Monit itself needs some management. A new module mediates the
interaction with monit so other modules don't need to understand the
configuration file format or how to add and remove themselves. A
single config file is shared between all services, with blocks
dedicated to each service. Two new commands start and stop the
monitoring daemon: admin monit start and admin monit stop. After
starting the daemon, starting services or templates that are
configured to use monit just rewrite the configuration and send a
signal to the daemon.

As an example, we also convert the melville template to use monit.