Filter use of cmake_with_tools.
There are some cases where cmake_with_tools can be detrimental. This
adds a filter for one of those (and a location where other filters
could be added). If you have NFS for home directories, then you're
wasting a lot of bandwidth to use ccache, which is likely
counterproductive. If we find an NFS mount, don't bother to use
CCache. This is very conservative, but since NFS is very common for
home directories, it's worth being conservative.